Remove handling specific libraries
From now on, pass everything what is needed via arguments (or CLI
interface). This allows to use footprints which were not explicitly
added so the version of kle2netlist does not have to be updated
constantly). There are drawbacks though:
- no ISO specific switch added anymore (don't want to add yet another
  argument for `--iso-switch-footprint`)
- fixed list of key widths - some libs might not have dedicated
  footprints for all of them

Both of these points could be solved by some footprints directories
searching but this is not implemented yet.